requirements.txt
chromadb>=1.4.0
geocoder>=1.38.1
googlemaps>=4.10.0
google-cloud-speech>=2.34.0
numpy>=2.4.0
ollama>=0.6.1
piper>=0.14.5
piper_tts>=1.3.0
protobuf>=6.33.2
pushbullet.py>=0.12.0
pvporcupine==3.0.0
pvrecorder>=1.2.7
PyAudio>=0.2.14
Requests>=2.32.5
sounddevice>=0.5.3
soundfile>=0.13.1
spacy>=3.8.11
spotipy>=2.25.2
textblob>=0.15.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0